Swapped in new tests
