+++ src/pyprop8/tests.py
@@ -1,165 +1,11 @@
-import numpy as np
 import pyprop8 as pp
-from pyprop8.utils import stf_trapezoidal, rtf2xyz, make_moment_tensor
-from collections import defaultdict
+from pyprop8.utils import stf_trapezoidal,make_moment_tensor,rtf2xyz
+import numpy as np
 
-
-def test_spatial_derivatives(
-    model,
-    source,
-    stations,
-    nt=257,
-    dt=0.5,
-    alpha=0.023,
-    pad_frac=1,
-    derivatives=None,
-    delta=1e-4,
-    delta_scales={"depth": 1, "phi": 1e-4, "r": 1},
-    source_time_function=None,
-    xyz=True,
-):
-    """
-    Compute finite-difference derivatives and compare with the 'analytic'
-    derivatives generated by pyprop8. All arguments are as required by
-    `pyprop8.compute_seismograms()` except:
-
-    delta        - float, perturbation for finite-difference calculation
-    delta_scales - dict, containing keys that correspond to the keywords
-                   available when creating a `pyprop8.DerivativeSwitches()`
-                   object, defining a parameter-specific rescaling for `delta`
-                   to account for the fact that parameters are specified in
-                   dimensionalised units. If key is not found, 'delta' is used
-                   unmodified.
-
-
-    """
-    print("Testing spatial derivatives against finite difference calculation...")
-    if source.nsources > 1:
-        raise ValueError("test_derivatives requires a single force/mechanism")
-    if source_time_function is None:
-        source_time_function = lambda w: stf_trapezoidal(w, 3, 6)
-    if derivatives is None:
-        derivatives = pp.DerivativeSwitches(r=True, phi=True, z=True)
-    if type(delta_scales) is not defaultdict:
-        # Make delta_scales have a default value of '1'
-        delta_scales = defaultdict(lambda: 1, **delta_scales)
-
-    nDimSta = stations.nDim
-    if derivatives.nderivs == 0:
-        raise ValueError("All derivatives are 'switched off'!")
-    elif derivatives.nderivs > 1:
-        nDimDerivs = 1
-    else:
-        nDimDerivs = 0
-    nDimChannels = 1
-    if nt > 1:
-        nDimTime = 1
-    else:
-        nDimTime = 0
-
-    # Function to handle slicing of derivative arrays regardless of shapes
-    deriv_comp = lambda sl: tuple(
-        nDimSta * [slice(None)]
-        + nDimDerivs * [sl]
-        + nDimChannels * [slice(None)]
-        + nDimTime * [slice(None)]
-    )
-
-    tt, seis0, drv = pp.compute_seismograms(
-        model,
-        source,
-        stations,
-        nt,
-        dt,
-        alpha,
-        pad_frac=pad_frac,
-        derivatives=derivatives,
-        source_time_function=source_time_function,
-        xyz=xyz,
-    )
-    assert len(seis0.shape) == nDimSta + nDimChannels + nDimTime
-    assert len(drv.shape) == nDimSta + nDimChannels + nDimTime + nDimDerivs
-    drverrs = []
-
-    if derivatives.r:
-        sta_pert = stations.copy()
-        sta_pert.rmin += delta * delta_scales["r"]
-        sta_pert.rmax += delta * delta_scales["r"]
-        tt, seis = pp.compute_seismograms(
-            model,
-            source,
-            sta_pert,
-            nt,
-            dt,
-            alpha,
-            pad_frac=pad_frac,
-            derivatives=None,
-            source_time_function=source_time_function,
-            xyz=xyz,
-        )
-        fd = (seis - seis0) / (delta * delta_scales["r"])
-        err = drv[deriv_comp(derivatives.i_r)] - fd
-        if nDimTime == 0:
-            norm = abs(drv[deriv_comp(derivatives.i_r)])
-        else:
-            norm = abs(drv[deriv_comp(derivatives.i_r)]).max(-1)
-        maxerr = (abs(err) / norm.reshape(*norm.shape, 1)).max()
-        drverrs += [maxerr]
-        print("Max error, 'r' derivative:   %f%%" % (100 * maxerr))
-    if derivatives.phi:
-        sta_pert = stations.copy()
-        sta_pert.phimin += delta * delta_scales["phi"]
-        sta_pert.phimax += delta * delta_scales["phi"]
-        tt, seis = pp.compute_seismograms(
-            model,
-            source,
-            sta_pert,
-            nt,
-            dt,
-            alpha,
-            pad_frac=pad_frac,
-            derivatives=None,
-            source_time_function=source_time_function,
-            xyz=xyz,
-        )
-        fd = (seis - seis0) * (180 / np.pi) / (delta * delta_scales["phi"])
-        err = drv[deriv_comp(derivatives.i_phi)] - fd
-        if nDimTime == 0:
-            norm = abs(drv[deriv_comp(derivatives.i_phi)])
-        else:
-            norm = abs(drv[deriv_comp(derivatives.i_phi)]).max(-1)
-        maxerr = (abs(err) / norm.reshape(*norm.shape, 1)).max()
-        drverrs += [maxerr]
-        print("Max error, 'phi' derivative: %f%%" % (100 * maxerr))
-    if derivatives.z:
-        source_pert = source.copy()
-        source_pert.dep -= delta * delta_scales["depth"]
-        tt, seis = pp.compute_seismograms(
-            model,
-            source_pert,
-            stations,
-            nt,
-            dt,
-            alpha,
-            pad_frac=pad_frac,
-            derivatives=None,
-            source_time_function=source_time_function,
-            xyz=xyz,
-        )
-        fd = (seis - seis0) / delta
-        err = drv[deriv_comp(derivatives.i_z)] - fd
-        if nDimTime == 0:
-            norm = abs(drv[deriv_comp(derivatives.i_z)])
-        else:
-            norm = abs(drv[deriv_comp(derivatives.i_z)]).max(-1)
-        maxerr = (abs(err) / norm.reshape(*norm.shape, 1)).max()
-        drverrs += [maxerr]
-        print("Max error, 'depth' derivative: %f%%" % (100 * maxerr))
-    return drverrs
-
-
-if __name__ == "__main__":
-    stations = pp.RegularlyDistributedReceivers(10, 150, 5, 0, 360, 8, depth=5)
+def tests():
+    print("Running tests. Using `pyprop8` from: %s"%pp.__file__)
+    print("")
+    print(" 1. Creating objects")
     model = pp.LayeredStructureModel(
         [
             (3.0, 1.8, 0.0, 1.02),
@@ -169,20 +15,157 @@ if __name__ == "__main__":
             (np.inf, 8.0, 4.56, 3.34),
         ]
     )
+
     source = pp.PointSource(
         0,
         0,
         20,
-        rtf2xyz(make_moment_tensor(340, 90, 0, 2.4e8, 0, 0)),
+        rtf2xyz(make_moment_tensor(340, 70, 20, 2.4e8, 0, 0)),
         np.zeros([3, 1]),
         0,
     )
-    errs = test_spatial_derivatives(model, source, stations)
-    if max(errs) > 1e-4:
-        print("")
-        print(
-            "*** Warning: Mismatch between analytic and finite-difference derivatives? ***"
-        )
-        print("")
-    else:
-        print("Analytic derivatives agree with finite-difference approximation")
+    stations = pp.RegularlyDistributedReceivers(30, 100, 7, 0, 360, 10, depth=3).asListOfReceivers()
+
+    derivs = pp.DerivativeSwitches(x=True, y=True, z=True)
+
+    source_time_function = lambda w: stf_trapezoidal(w, 3, 6)
+
+    print(" 2. Computing seismograms...")
+
+    nt=33 #257
+    dt=0.5
+    alpha=0.023
+    pad_frac=1
+
+    tt, seis0, drv = pp.compute_seismograms(
+        model,
+        source,
+        stations,
+        nt,
+        dt,
+        alpha,
+        pad_frac=pad_frac,
+        derivatives=derivs,
+        source_time_function=source_time_function,
+        xyz=True,
+    )
+
+    epsilon = 1e-4
+    print(" 3. Comparing with finite-difference derivatives.")
+    print("    Using finite-difference perturbation eps = %.2f metres"%(epsilon*1000))
+    print("    a. Perturbing source in x.")
+
+
+    source_x = source.copy()
+    source_x.x += epsilon
+    
+    tt, seis_x = pp.compute_seismograms(
+        model,
+        source_x,
+        stations,
+        nt,
+        dt,
+        alpha,
+        pad_frac=pad_frac,
+        derivatives=None,
+        source_time_function=source_time_function,
+        xyz=True,
+    )
+
+    fd = (seis_x - seis0)/epsilon # finite difference estimate
+    max_x = abs(drv[:,derivs.i_x,:,:]).max(-1).reshape(stations.nstations,3,1) # Maximum absolute value of trace
+    perc_err_x = 100*(abs(drv[:,derivs.i_x,:,:]-fd)/max_x) #For each trace, calculate finite difference 'error' as percentage of trace amplitude
+    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%perc_err_x.max())
+
+    print("    b. Perturbing source in y.")
+    source_y = source.copy()
+    source_y.y += epsilon
+    
+    tt, seis_y = pp.compute_seismograms(
+        model,
+        source_y,
+        stations,
+        nt,
+        dt,
+        alpha,
+        pad_frac=pad_frac,
+        derivatives=None,
+        source_time_function=source_time_function,
+        xyz=True,
+    )
+
+    fd = (seis_y - seis0)/epsilon # finite difference estimate
+    max_y = abs(drv[:,derivs.i_y,:,:]).max(-1).reshape(stations.nstations,3,1) # Maximum absolute value of trace
+    perc_err_y = 100*(abs(drv[:,derivs.i_y,:,:]-fd)/max_y) #For each trace, calculate finite difference 'error' as percentage of trace amplitude
+    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%perc_err_y.max())
+    
+
+    print("    c. Perturbing source in z.")
+    source_z = source.copy()
+    source_z.dep -= epsilon # coordinate system is z-up so a positive epsilon in z is a *reduction* in source depth
+    
+    tt, seis_z = pp.compute_seismograms(
+        model,
+        source_z,
+        stations,
+        nt,
+        dt,
+        alpha,
+        pad_frac=pad_frac,
+        derivatives=None,
+        source_time_function=source_time_function,
+        xyz=True,
+    )
+
+    fd = (seis_z - seis0)/epsilon # finite difference estimate
+    max_z = abs(drv[:,derivs.i_z,:,:]).max(-1).reshape(stations.nstations,3,1) # Maximum absolute value of trace
+    perc_err_z = 100*(abs(drv[:,derivs.i_z,:,:]-fd)/max_z) #For each trace, calculate finite difference 'error' as percentage of trace amplitude
+    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%perc_err_z.max())
+
+    print(" 4. Computing static displacement field")
+    stat0, drv = pp.compute_static(
+        model,
+        source,
+        stations,
+        derivatives=derivs,
+    )
+    
+    print(" 5. Comparing with finite-difference derivatives")
+    print("    a. Perturbing source in x.")
+   
+    stat_x = pp.compute_static(
+        model,
+        source_x,
+        stations,
+    )
+
+    fd = (stat_x - stat0)/epsilon
+    max_x  = abs(drv[:,derivs.i_x,:]).max(0).reshape(1,3)
+    perc_err_x = 100*abs(drv[:,derivs.i_x,:]-fd)/max_x
+    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%(perc_err_x.max()))
+    print("    b. Perturbing source in y.")
+   
+    stat_y = pp.compute_static(
+        model,
+        source_y,
+        stations,
+    )
+    fd = (stat_y - stat0)/epsilon
+
+    max_y  = abs(drv[:,derivs.i_y,:]).max(0).reshape(1,3)
+    perc_err_y = 100*abs(drv[:,derivs.i_y,:]-fd)/max_y
+    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%(perc_err_y.max()))
+    print("    c. Perturbing source in z.")
+   
+    stat_z = pp.compute_static(
+        model,
+        source_z,
+        stations,
+    )
+    fd = (stat_z - stat0)/epsilon
+    max_z  = abs(drv[:,derivs.i_z,:]).max(0).reshape(1,3)
+    perc_err_z = 100*abs(drv[:,derivs.i_z,:]-fd)/max_z
+    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%(perc_err_z.max()))
+
+if __name__ == "__main__":
+    tests()
